requirements.txt
openai-whisper
torch
fastapi
uvicorn[standard]
python-multipart
numpy
